+++ ricaun.RevitTest.Command/Command/RunCommand.cs
@@ -1,4 +1,5 @@
-﻿using ricaun.RevitTest.Command.Extensions;
+﻿using ricaun.NUnit;
+using ricaun.RevitTest.Command.Extensions;
 using ricaun.RevitTest.Command.Utils;
 using System;
 using System.IO;
@@ -30,7 +31,7 @@ namespace ricaun.RevitTest.Command
 
         private void Initialize()
         {
-            NUnit.TestEngine.Initialize(out string init);
+            TestEngine.Initialize(out string init);
 
             Log.WriteLine();
             Log.WriteLine($"{AppUtils.GetInfo()} [{init}]");
@@ -77,8 +78,8 @@ namespace ricaun.RevitTest.Command
                 outputAction,
                 options.RevitLanguage,
                 options.ForceToOpen,
-                options.ForceToWait,
                 options.ForceToClose,
+                options.Timeout,
                 options.Test.ToArray());
 
             return false;