+++ ricaun.RevitTest.Console/Program.cs
@@ -7,13 +7,6 @@ namespace ricaun.RevitTest.Console
         static void Main(string[] args)
         {
             var result = RunTest.ParseArguments<Revit.RevitRunTestService>(args);
-
-#if DEBUG
-            if (!result)
-            {
-                Revit.Utils.RevitDebugUtils.ProcessServerSelect();
-            }
-#endif
         }
     }
 }